% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInstallPkg.R
\name{check_install}
\alias{check_install}
\title{Check and Install R Package}
\usage{
check_install(package_name)
}
\arguments{
\item{package_name}{A character string specifying the name of the package to be checked and installed.}
}
\value{
TRUE if the package is already installed or can be installed. FALSE otherwise
}
\description{
This function checks if an R package is installed, and if not, attempts to install
it using either the standard CRAN repository or the Bioconductor repository.
}
\examples{
\donttest{
# Check and install "dplyr" package
check_install("dplyr")
}

}
