% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFilenames.R
\name{extractFilenames}
\alias{extractFilenames}
\title{Extract file names from user prompt}
\usage{
extractFilenames(text)
}
\arguments{
\item{text}{user prompt}
}
\description{
This function extracts file names from the user prompt. Current filenames that
are supported by this function are *.txt, *.tsv, *.csv *.xls, *.xlsx,
*.bed, *.bigWig, *.bw and *.bigBed. Other filenames will not be extracted.
If no filenames are found, the function will return NA.
}
\examples{
\dontrun{
extractFilenames(text=user_prompt)
}
}
