% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileHeaderPrompt.R
\name{fileHeaderPrompt}
\alias{fileHeaderPrompt}
\title{Extract file headers from files in prompt}
\usage{
fileHeaderPrompt(filenames)
}
\arguments{
\item{filenames}{list containing file names.}
}
\value{
A string containing the file headers of the files in "filenames" list
}
\description{
This function extracts file headers from
files. Recommended is to use this function
with the results from \code{\link{extractFilenames}}
}
\examples{
\dontrun{
prompt <- "how do I perform PCA on data in a file called test.txt?"
filenames <- extractFilenames(prompt)
fileHeaderPrompt(filenames)
}
}
