% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promptContext.R
\name{promptContext}
\alias{promptContext}
\title{Predefined prompt contexts for prompt engineering}
\usage{
promptContext(type = "simple")
}
\arguments{
\item{type}{specifies the type of context you wish to be returned.
Valid options are "simple", "actAs", "CoT" and "rbionfoExp"}
}
\value{
A string holding the predefined context.
}
\description{
This function holds various predefined prompt
contexts that can be used for prompt engineering.
}
