% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupAgent.R
\name{setupAgent}
\alias{setupAgent}
\title{set up an online LLM API for subsequent tasks}
\usage{
setupAgent(
  name = c("openai", "replicate"),
  type = NULL,
  model,
  ai_api_key = Sys.getenv("AI_API_KEY")
)
}
\arguments{
\item{name}{Name of the API you want to use. Currently supported APIs are "openai" and "replicate"}

\item{type}{Specify type of model (chat or completion). This parameter only needs to be specified when using 'openai}

\item{model}{LLM model you wish to use.}

\item{ai_api_key}{personal API key for accessing LLM}
}
\description{
This function sets up an large language model API for tasks.
}
\examples{
\dontrun{
myAgent <- setupAgent(name="openai",type="chat",model="gpt-4",ai_api_key=Sys.getenv("AI_API_KEY"))

myAgent <- setupAgent(name="replicate",type=NULL,model="llama-2-70b-chat",ai_api_key=Sys.getenv("AI_API_KEY"))
}
}
