% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupAgent.R
\name{setupopenaiAgent}
\alias{setupopenaiAgent}
\title{setup the online large language model (LLM) API for subsequent tasks}
\usage{
setupopenaiAgent(
  model,
  type = c("chat", "completion"),
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL
)
}
\description{
setup the online large language model (LLM) API for subsequent tasks
}
