% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupAgent.R
\name{setupTestAgent}
\alias{setupTestAgent}
\title{setup a test agent to be used for test prompting
should return a random set of responses}
\usage{
setupTestAgent(agentName = "testAgent")
}
\description{
setup a test agent to be used for test prompting
should return a random set of responses
}
