% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_code_blocks.R
\name{clean_code_blocks}
\alias{clean_code_blocks}
\title{Clean code blocks returned by the agent}
\usage{
clean_code_blocks(response)
}
\arguments{
\item{response}{response recieved by the agent}
}
\description{
This function cleans up the code blocks which
are returned by the agent to ensure code blocks
can run.
}
\examples{
\dontrun{
clean_code <- clean_code_blocks(response)
}
}
