% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFilenames.R
\name{extractFilenames}
\alias{extractFilenames}
\title{Extract file names from user prompt}
\usage{
extractFilenames(text)
}
\arguments{
\item{text}{user prompt}
}
\description{
This function extracts file names from the user prompt
}
\examples{
\dontrun{
extractFilenames(text=user_prompt)
}
}
