% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInstallPkg.R
\name{extractInstallPkg}
\alias{extractInstallPkg}
\title{extract package names and install them}
\usage{
extractInstallPkg(code)
}
\arguments{
\item{code}{code block returned by the agent.}
}
\description{
This function extracts all package names that
are needed to run the code returned by the agent
and installs them as needed.
}
\examples{
\dontrun{
# Check code for packages that need installing
extractInstallPkg(code)
}

}
