% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileHeaderPrompt.R
\name{fileHeaderPrompt}
\alias{fileHeaderPrompt}
\title{Extract file headers from files in prompt}
\usage{
fileHeaderPrompt(filenames)
}
\arguments{
\item{filenames}{list containing filenames mentioned in prompt}
}
\description{
This function extracts file headers from
files mentioned in user prompts.
}
\examples{
\dontrun{
fileHeaderPrompt(filenames)
}
}
