% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileHeaderPrompt.R
\name{fileHeaderPrompt}
\alias{fileHeaderPrompt}
\title{Extract file headers from files in prompt}
\usage{
fileHeaderPrompt(filenames)
}
\arguments{
\item{filenames}{list containing filenames.}
}
\description{
This function extracts file headers from
files. Recommended is to use this function
with the results from \code{\link{extractFilenames}}
}
\examples{
\dontrun{
fileHeaderPrompt(filenames)
}
}
