% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupAgent.R
\name{setupAgent}
\alias{setupAgent}
\title{setup any online  large language model (LLM) API for subsequent tasks}
\usage{
setupAgent(URL, task, model, ai_api_key, authorization_name)
}
\arguments{
\item{URL}{base url for LLM you wish to use}

\item{task}{specific task LLM needs to complete. Syntax depends on LLM URL used.}

\item{model}{model you wish to use}

\item{ai_api_key}{personal API key for accessing LLM}

\item{authorization_name}{name of authorization type to be parsed to website. Depends on URL used.}
}
\description{
This function sets up any online large language model API for tasks.
}
\examples{
\dontrun{
myAgent <- setupAgent (URL="https://api.replicate.com/v1/", task = "predictions",
model="02e509c789964a7ea8736978a43525956ef40397be9033abf9fd2badfe68c9e3",
ai_api_key=Sys.getenv("API_KEY"), authorization_name = "Token")
myAgent <- setupAgent (URL="https://api.openai.com/v1/", task = "chat/completions",
model="gtp-4",
ai_api_key=Sys.getenv("API_KEY"), authorization_name = "Bearer")
}
}
