% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupAgent.R
\name{setupopenaiAgent}
\alias{setupopenaiAgent}
\title{Setting up LLM agent}
\usage{
setupopenaiAgent(
  model,
  type = c("chat", "completion"),
  openai_api_key = Sys.getenv("OPENAI_API_KEY"),
  openai_organization = NULL
)
}
\arguments{
\item{model}{Specifies LLM to be used}

\item{type}{Specifies type of model (chat or completion)}

\item{openai_api_key}{openAI API key}

\item{openai_organization}{openAI organization}
}
\description{
This function is used for the
setup of the online large language model (LLM) API for subsequent tasks
}
